# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TechLane API
  version: 1.0.0
  description: Shared REST API for TechLane ops web, Android, and portals.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /part-requests:
    post:
      summary: Request part for repair
      operationId: createPartRequest
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartRequest"
  /part-requests/{id}/approve:
    post:
      summary: Approve part request and issue auth code
      operationId: approvePartRequest
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                supplier_id:
                  type: string
                  format: uuid
                unit_cost:
                  type: number
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierIssue"
  /supplier-issues/{id}/collect:
    post:
      summary: Record supplier part collection
      operationId: collectSupplierIssue
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [auth_code]
              properties:
                auth_code:
                  type: string
      responses:
        "200":
          description: Collected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierIssue"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema:
        type: string
        format: uuid
  schemas:
...
    CreatePartRequest:
      type: object
      required: [repair_job_id, description]
      properties:
        repair_job_id:
          type: string
          format: uuid
        variant_id:
          type: string
          format: uuid
        description:
          type: string
        quantity:
          type: integer
          default: 1
    PartRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        repair_job_id:
          type: string
          format: uuid
        status:
          type: string
        description:
          type: string
    SupplierIssue:
      type: object
      properties:
        id:
          type: string
          format: uuid
        part_request_id:
          type: string
          format: uuid
        repair_job_id:
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        auth_code:
          type: string
        status:
          type: string
        unit_cost:
          type: number
        collected_at:
          type: string
          format: date-time
        reconciliation_status:
          type: string
...